move concurrency utils to their own pkg
