+++ internal/concurrency/concurrency.go
@@ -1,4 +1,4 @@
-package utils
+package concurrency
 
 import (
 	"context"
